+++ CHANGELOG.md
@@ -2,5 +2,9 @@
 Systemlessly replaces emoji font with WhatsApp Emoji 
 
 ## Changelog
+v1502
+- Replace font in more places
+- Improved install messages
+
 v1501
 - Initial release using [this font file](https://github.com/dmlls/whatsapp-emoji-linux/blob/main/png/160/emoji_u1f3f3_200d_1f308.png)